# Release v0.4.3.1b 2025-12-30
## New Features
1. make replication shard count agnostic (v0.4.3.0b 2025-12-28)
2. add debug replication mode (v0.4.3.0b 2025-12-28)

## Fixes
1. fix string push along with `CLIENT INFO` (v0.4.3.0b 2025-12-29)
2. hash test did not test hash (v0.4.3.0b 2025-12-29)
3. remove unused code (v0.4.3.1b 2025-12-30)
4. fix incr/decr for doubles

+++ src/swig_api.cpp
@@ -299,7 +299,7 @@ Value KeyValue::max() const {
 }
 
 void KeyValue::incr(const std::string& key, double by) {
-    std::string b = std::to_string(by);
+    std::string b = std::to_string((int64_t)by);
     params = {"INCRBY",key, b};
 
     int r = sc.call(params, ::INCRBY);
@@ -307,7 +307,7 @@ void KeyValue::incr(const std::string& key, double by) {
 }
 
 void KeyValue::decr(const std::string& key, double by) {
-    std::string b = std::to_string(by);
+    std::string b = std::to_string((int64_t)by);
     params = {"DECRBY", key,b};
 
     int r = sc.call(params, ::DECRBY);
